Create game logic and structure
I noticed that in the choices array ["Rock", "Paper", "Scissors"],
choices[2] is the same as choices[-1], so choice[x] always beats
choice[x-1]. The conditional block uses that logic to determine the
winner. Feels like the most elegant way to solve it! I thought about
hardcoding with && statements or using an array of arrays, but the -1
thing is simpler and more straightforward. Need to fix the prompt to
handle type errors.